sleep: wait for firmware load timeout
The firmware loader seems to be buggy and can break system suspend
if a load is in progress while the system suspends.

Since this test case may run immediatly after boot, there's a strong
chance that the firmware loader is still running/waiting, so wait for
it to timeout before attempting to suspend.

Signed-off-by: Kevin Hilman <khilman@baylibre.com>